requirements.txt
python-dotenv
openai 
modelscope